Process EM38 data

This function processes the outputs of n38_decode into a useable end product.

em38_decode(n38_decoded = NULL, hdop_filter = 3, time_filter = 5,
  fix_filter = NULL)

Arguments

n38_decoded
Nested list output by n38_decode.

hdop_filter
Numeric, discard GPS data where the Horizontal Dilution of Precision is greater than this number. Defaults to 3 metres. Set to NULL to keep all readings.

time_filter
Numeric, discard readings taken more than n seconds after the last acceptable GPS reading. Set to 2-3x GPS aquisition frequency. Defaults to 5 seconds. Set to NULL to keep all readings.

fix_filter
Select readings with a minimum quality of GPS fix. Options are:

Autonomous GPS fix.

DGPS fix, using a local DGPS base station or correction service such as WAAS or EGNOS.

Pulse per second (PPS) fix.

real-time kinematic (RTK) fix.

RTK float fix.

estimated (dead reckoning)

manual input mode

simulation mode

WAAS fix.

To filter on multiple options, supply a vector e.g. c(2, 9). Defaults to NULL.

Value

A two element list containing file header data, and a list of processed survey lines. For each survey line, an `sf` data frame with sfc_POINT geometry is returned where valid GPS data exists. If instrument readings without valid GPS data are present, a data frame is returned. Otherwise, a string is returned explaining the failure to process.

Examples

data('n38_demo')
n38_chunks  <- n38_chunk(n38_demo)
n38_decoded <- n38_decode(n38_chunks)
demo_survey <- em38_decode(n38_decoded, 3)