Import and convert N38 logfiles

This is a wrapper function that processes a raw on-disk N38 file into useable data.

em38_from_file(
  path = NULL,
  hdop_filter = 3,
  time_filter = 5,
  fix_filter = NULL
)

Arguments

path
A file path pointing to a valid *.N38 file, produced by a Geonics EM38-MK2 conductivity sensor connected to an Allegra CX or Archer datalogger (and optionally, a GPS device).

hdop_filter
Numeric, discard GPS data where the Horizontal Dilution of Precision is greater than this number. Defaults to 3 metres. Set to NULL to keep all readings.

time_filter
Numeric, discard readings taken more than n seconds after the last acceptable GPS reading. Set to 2-3x GPS aquisition frequency. Defaults to 5 seconds. Set to NULL to keep all readings.

fix_filter
Select readings with a minimum quality of GPS fix. Options are:

Autonomous GPS fix.

DGPS fix, using a local DGPS base station or correction service such as WAAS or EGNOS.

Pulse per second (PPS) fix.

real-time kinematic (RTK) fix.

RTK float fix.

estimated (dead reckoning)

manual input mode

simulation mode

WAAS fix.

To filter on multiple options, supply a vector e.g. c(2, 9). Defaults to NULL.

Value

A two element list containing file header data, and a list of processed survey lines. For each survey line, an `sf` data frame with sfc_POINT geometry is returned where valid GPS data exists. If instrument readings without valid GPS data are present, a data frame is returned. Otherwise, a string is returned explaining the failure to process.

Examples

demo_survey <-
em38_from_file(path = system.file("extdata", "em38_demo.N38", package = "em38"),
              hdop_filter = 3)